Fix: controller test suite
