+++ controllers/addon_controller_test.go
@@ -7,29 +7,34 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 
-	wfv1api "github.com/argoproj/argo-workflows/v3/pkg/apis/workflow/v1alpha1"
 	"github.com/keikoproj/addon-manager/api/addon/v1alpha1"
 )
 
 var (
-	addonNamespace = "default"
-	addonName      = "cluster-autoscaler"
-	addonKey       = types.NamespacedName{Name: addonName, Namespace: addonNamespace}
+	addonKey = types.NamespacedName{Name: addonName, Namespace: addonNamespace}
 )
 
-const timeout = time.Second * 5
+var fetched *v1alpha1.Addon
 
-var instance *v1alpha1.Addon
-var updated *v1alpha1.Addon
-var fetchedAddon *v1alpha1.Addon
-var addonController *Controller
+const timeout = time.Second * 5
 
 var _ = Describe("AddonController", func() {
 
 	Describe("Addon CR can be reconciled", func() {
+		var instance *v1alpha1.Addon
+		var wfv1 = &unstructured.Unstructured{}
+		wfv1.SetGroupVersionKind(schema.GroupVersionKind{
+			Kind:    "Workflow",
+			Group:   "argoproj.io",
+			Version: "v1alpha1",
+		})
+
 		It("instance should be parsable", func() {
 			addonYaml, err := ioutil.ReadFile("./tests/clusterautoscaler.yaml")
 			Expect(err).ToNot(HaveOccurred())
@@ -38,17 +43,14 @@ var _ = Describe("AddonController", func() {
 			Expect(err).ToNot(HaveOccurred())
 			Expect(instance).To(BeAssignableToTypeOf(&v1alpha1.Addon{}))
 			Expect(instance.GetName()).To(Equal(addonName))
-
 		})
 
 		It("instance should be reconciled", func() {
-			instance.SetNamespace(addonNamespace)
-			addonController = newController(instance)
-
 			processed := addonController.processNextItem(ctx)
 			Expect(processed).To(BeTrue())
-			fetchedAddon, err := addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Get(ctx, testAddon, metav1.GetOptions{})
-			Expect(err).To(BeNil())
+
+			fetchedAddon, err := addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Get(ctx, instance.Name, metav1.GetOptions{})
+			Expect(err).NotTo(HaveOccurred())
 			Expect(fetchedAddon).NotTo(BeNil())
 
 			By("Verify addon has been reconciled by checking for checksum status")
@@ -56,30 +58,41 @@ var _ = Describe("AddonController", func() {
 
 			By("Verify addon has finalizers added which means it's valid")
 			Expect(fetchedAddon.ObjectMeta.Finalizers).Should(Equal([]string{"delete.addonmgr.keikoproj.io"}))
+
 			oldCheckSum := instance.Status.Checksum
 			//Update instance params for checksum validation
 			instance.Spec.Params.Context.ClusterRegion = "us-east-2"
-			err = addonController.handleAddonUpdate(ctx, fetchedAddon)
-			Expect(err).To(BeNil())
+			//err = k8sClient.Update(context.TODO(), instance)
+			_, err = addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Update(ctx, fetchedAddon, metav1.UpdateOptions{})
+			if apierrors.IsInvalid(err) {
+				addonController.logger.Error(err, "failed to update object, got an invalid object error")
+				return
+			}
+			Expect(err).NotTo(HaveOccurred())
 
 			Eventually(func() error {
-				updated, err = addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Get(ctx, fetchedAddon.Name, metav1.GetOptions{})
-				Expect(err).To(BeNil())
-				Expect(updated).NotTo(BeNil())
-				Expect(updated.Status.Checksum).ShouldNot(BeIdenticalTo(oldCheckSum))
-				return nil
+				fetched, err = addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Get(ctx, instance.Name, metav1.GetOptions{})
+				if err != nil {
+					return err
+				}
+
+				if len(fetched.ObjectMeta.Finalizers) > 0 {
+					return nil
+				}
+				return fmt.Errorf("addon is not valid")
 			}, timeout).Should(Succeed())
 
 			By("Verify changing addon spec generates new checksum")
-			Expect(updated.Status.Checksum).ShouldNot(BeIdenticalTo(oldCheckSum))
+			Expect(fetched.Status.Checksum).ShouldNot(BeIdenticalTo(oldCheckSum))
 
 			By("Verify addon has workflows generated with new checksum name")
-			wfName := updated.GetFormattedWorkflowName(v1alpha1.Prereqs)
-			var fetchedwf *wfv1api.Workflow
+			wfName := fetched.GetFormattedWorkflowName(v1alpha1.Prereqs)
 			Eventually(func() error {
-				fetchedwf, err = addonController.wfcli.ArgoprojV1alpha1().Workflows(testNamespace).Get(ctx, wfName, metav1.GetOptions{})
-				Expect(err).To(BeNil())
-				Expect(fetchedwf.GetName()).Should(Equal(wfName))
+				wfv1, err := addonController.wfcli.ArgoprojV1alpha1().Workflows(testNamespace).Get(ctx, wfName, metav1.GetOptions{})
+				if err != nil {
+					return err
+				}
+				Expect(wfv1.GetName()).Should(Equal(wfName))
 				return nil
 			}, timeout).Should(Succeed())
 
@@ -93,23 +106,30 @@ var _ = Describe("AddonController", func() {
 
 		It("instance should be deleted w/ deleting state", func() {
 			By("Verify deleting instance should set Deleting state")
-			Expect(addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Delete(ctx, updated.Name, metav1.DeleteOptions{})).NotTo(HaveOccurred())
+			fetched.ObjectMeta.DeletionTimestamp = &metav1.Time{Time: metav1.Now().Time}
+			addonController.handleAddonDeletion(ctx, fetched)
 			Eventually(func() error {
-				_, err := addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Get(ctx, updated.Name, metav1.GetOptions{})
-				Expect(err).NotTo(BeNil())
+				updated, err := addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Get(ctx, fetched.Name, metav1.GetOptions{})
+				if err != nil {
+					return err
+				}
 
-				return nil
+				if updated.Status.Lifecycle.Installed == v1alpha1.Deleting {
+					return nil
+				}
+				return fmt.Errorf("addon is not being deleted")
 			}, timeout).Should(Succeed())
 
 			By("Verify delete workflow was generated")
-			wfName := instance.GetFormattedWorkflowName(v1alpha1.Delete)
+			wfName := fetched.GetFormattedWorkflowName(v1alpha1.Delete)
 			Eventually(func() error {
-				wf, err := addonController.wfcli.ArgoprojV1alpha1().Workflows(addonNamespace).Get(ctx, wfName, metav1.GetOptions{})
-				Expect(err).NotTo(BeNil())
-				Expect(wf).To(BeNil())
+				wfv1, err := addonController.wfcli.ArgoprojV1alpha1().Workflows(testNamespace).Get(ctx, wfName, metav1.GetOptions{})
+				if err != nil {
+					return err
+				}
+				Expect(wfv1.GetName()).Should(Equal(wfName))
 				return nil
 			}, timeout).Should(Succeed())
-
 		})
 
 		It("instance with dependencies should succeed", func() {
@@ -144,40 +164,37 @@ var _ = Describe("AddonController", func() {
 			}
 
 			By("Verify first addon-2 that depends on addon-1 is created and has validation failed state")
-			createdInstance2, err := addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Create(ctx, instance2, metav1.CreateOptions{})
+			_, err := addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Create(ctx, instance2, metav1.CreateOptions{})
 			Expect(err).To(BeNil())
-			Expect(createdInstance2).NotTo(BeNil())
 			err = addonController.handleAddonCreation(ctx, instance2)
 			Expect(err).NotTo(BeNil())
-
-			defer addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Delete(ctx, instance2.Name, metav1.DeleteOptions{})
 			var fetchedInstance2 *v1alpha1.Addon
 			Eventually(func() error {
 				fetchedInstance2, err = addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Get(ctx, instance2.Name, metav1.GetOptions{})
-				if err != nil || fetchedAddon == nil {
+				if err != nil {
 					return err
 				}
 
-				if fetchedInstance2.Status.Lifecycle.Installed == v1alpha1.ValidationFailed {
+				if fetchedInstance2.Status.Lifecycle.Installed == v1alpha1.DepNotInstalled {
 					return nil
 				}
-
 				return fmt.Errorf("addon-2 is not in validation failed state")
 			}, timeout).Should(Succeed())
 
 			By("Verify addon-1 is submitted and completes successfully")
-			createdInstance1, err := addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Create(ctx, instance, metav1.CreateOptions{})
+
+			_, err = addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Create(ctx, instance, metav1.CreateOptions{})
 			Expect(err).To(BeNil())
-			Expect(createdInstance1).NotTo(BeNil())
 			err = addonController.handleAddonCreation(ctx, instance)
 			Expect(err).To(BeNil())
 			defer addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Delete(ctx, instance.Name, metav1.DeleteOptions{})
-
 			Eventually(func() error {
-				fetchedInstance, err := addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Get(ctx, instance.Name, metav1.GetOptions{})
-				Expect(err).NotTo(HaveOccurred())
+				instance, err := addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Get(ctx, instance.Name, metav1.GetOptions{})
+				if err != nil {
+					return err
+				}
 
-				if fetchedInstance.Status.Lifecycle.Installed == v1alpha1.Succeeded {
+				if instance.Status.Lifecycle.Installed == v1alpha1.Succeeded {
 					return nil
 				}
 
@@ -188,15 +205,17 @@ var _ = Describe("AddonController", func() {
 			err = addonController.handleAddonUpdate(ctx, fetchedInstance2)
 			Expect(err).To(BeNil())
 			Eventually(func() error {
-				updatedInstance2, err := addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Get(ctx, instance2.Name, metav1.GetOptions{})
-				Expect(err).NotTo(HaveOccurred())
-				Expect(fetchedInstance2).NotTo(BeNil())
+				instance2, err := addonController.addoncli.AddonmgrV1alpha1().Addons(testNamespace).Get(ctx, instance2.Name, metav1.GetOptions{})
+				if err != nil {
+					return err
+				}
 
-				if updatedInstance2.Status.Lifecycle.Installed == v1alpha1.Succeeded {
+				if instance2.Status.Lifecycle.Installed == v1alpha1.Succeeded {
 					return nil
 				}
-				return fmt.Errorf("addon-2 is not valid after resolving dependencies.")
-			}, timeout).Should(Succeed())
+
+				return fmt.Errorf("addon-2 is not valid")
+			}, timeout*10).Should(Succeed())
 		})
 
 	})
