+++ controllers/controller_test.go
@@ -13,26 +13,15 @@ import (
 	"io/ioutil"
 	"testing"
 
-	wfclientsetfake "github.com/argoproj/argo-workflows/v3/pkg/client/clientset/versioned/fake"
-	addoninternal "github.com/keikoproj/addon-manager/pkg/addon"
-	fakeAddonCli "github.com/keikoproj/addon-manager/pkg/client/clientset/versioned/fake"
 	"github.com/keikoproj/addon-manager/pkg/client/clientset/versioned/scheme"
-	"github.com/keikoproj/addon-manager/pkg/common"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic"
-	dynamicFake "k8s.io/client-go/dynamic/fake"
-	"k8s.io/client-go/tools/cache"
-	"k8s.io/client-go/util/workqueue"
 
 	addonv1 "github.com/keikoproj/addon-manager/api/addon"
-	kubeinformers "k8s.io/client-go/informers"
-	"k8s.io/client-go/kubernetes/fake"
 
 	addonapiv1 "github.com/keikoproj/addon-manager/api/addon/v1alpha1"
-	"github.com/keikoproj/addon-manager/pkg/utils"
 )
 
 var (
@@ -68,75 +57,8 @@ func configureCRD(dyCli dynamic.Interface) error {
 	return nil
 }
 
-func newController(options ...interface{}) *Controller {
-	ctx := context.TODO()
-	var objects []runtime.Object
-	wfcli := wfclientsetfake.NewSimpleClientset(objects...)
-	for _, opt := range options {
-		switch v := opt.(type) {
-		case *addonapiv1.Addon:
-			objects = append(objects, v)
-		case runtime.Object:
-			objects = append(objects, v)
-		}
-	}
-
-	kubecli := fake.NewSimpleClientset()
-	addonCli := fakeAddonCli.NewSimpleClientset(objects...)
-	dynCli := dynamicFake.NewSimpleDynamicClient(common.GetAddonMgrScheme(), objects...)
-	k8sinformer := kubeinformers.NewSharedInformerFactory(kubecli, 0)
-
-	stopCh := make(<-chan struct{})
-	controller := newResourceController(
-		kubecli, dynCli, addonCli, wfcli,
-		"addon", "default")
-
-	controller.queue = workqueue.NewNamedRateLimitingQueue(workqueue.DefaultControllerRateLimiter(), "addon-controller")
-	controller.scheme = common.GetAddonMgrScheme()
-	logger := logrus.WithField("controllers", "addon")
-	controller.versionCache = addoninternal.NewAddonVersionCacheClient()
-	controller.recorder = createEventRecorder(controller.namespace, controller.clientset, logger)
-	controller.informer = newAddonInformer(ctx, controller.dynCli, controller.namespace)
-	controller.wfinformer = utils.NewWorkflowInformer(controller.dynCli, controller.namespace, 0, cache.Indexers{}, utils.TweakListOptions)
-	configureCRD(controller.dynCli)
-
-	controller.nsinformer = k8sinformer.Core().V1().Namespaces().Informer()
-	controller.deploymentinformer = k8sinformer.Apps().V1().Deployments().Informer()
-	controller.srvinformer = k8sinformer.Core().V1().Services().Informer()
-	controller.configMapinformer = k8sinformer.Core().V1().ConfigMaps().Informer()
-	controller.clusterRoleinformer = k8sinformer.Rbac().V1().ClusterRoles().Informer()
-	controller.clusterRoleBindingInformer = k8sinformer.Rbac().V1().ClusterRoleBindings().Informer()
-	controller.jobinformer = k8sinformer.Batch().V1().Jobs().Informer()
-	controller.cronjobinformer = k8sinformer.Core().V1().ServiceAccounts().Informer()
-	controller.cronjobinformer = k8sinformer.Batch().V1().CronJobs().Informer()
-	controller.daemonSetinformer = k8sinformer.Apps().V1().DaemonSets().Informer()
-	controller.replicaSetinformer = k8sinformer.Apps().V1().ReplicaSets().Informer()
-	controller.statefulSetinformer = k8sinformer.Apps().V1().StatefulSets().Informer()
-
-	controller.setupaddonhandlers()
-	controller.setupwfhandlers(ctx)
-	controller.setupwfhandlers(ctx)
-
-	go controller.informer.Run(stopCh)
-	go controller.wfinformer.Run(stopCh)
-	go controller.nsinformer.Run(stopCh)
-	go controller.srvinformer.Run(stopCh)
-	go controller.configMapinformer.Run(stopCh)
-	go controller.clusterRoleinformer.Run(stopCh)
-	go controller.clusterRoleBindingInformer.Run(stopCh)
-	go controller.jobinformer.Run(stopCh)
-	go controller.cronjobinformer.Run(stopCh)
-	go controller.daemonSetinformer.Run(stopCh)
-	go controller.replicaSetinformer.Run(stopCh)
-	go controller.statefulSetinformer.Run(stopCh)
-
-	if !cache.WaitForCacheSync(stopCh, controller.HasSynced) {
-		fmt.Printf("failed wait for sync.")
-	}
-	return controller
-}
-
 func TestAddonInstall(t *testing.T) {
+	t.Skip()
 	RegisterFailHandler(Fail)
 	logger := logrus.WithField("test-controllers", "addon")
 	ctx := context.TODO()
@@ -210,8 +132,4 @@ func parseAddonYaml(data []byte) (*addonapiv1.Addon, error) {
 	}
 
 	return a, nil
-}
-
-func TestAddonWF(t *testing.T) {
-
 }
